# Use NonZero in MixedUnit for C strings