--- a/.github/workflows/CI.yml
+++ b/.github/workflows/CI.yml
@@ -11,13 +11,13 @@ env:
   RUST_BACKTRACE: 1
 
 jobs:
-  stable-checks:
+  beta-checks:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-      - uses: dtolnay/rust-toolchain@stable
+      - uses: dtolnay/rust-toolchain@beta
         with:
-          toolchain: stable
+          toolchain: beta
           components: rustfmt
       - run: cargo fmt -- --check
       - run: cargo test