--- a/Cargo.toml
+++ b/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 description = "Provides code to unescape string literals"
 license = "Apache-2.0 OR MIT"
 repository = "https://github.com/rust-lang/literal-escaper"
+rust-version = "1.89" # for NonZero<char>
 
 [dependencies]
 std = { version = '1.0.0', optional = true, package = 'rustc-std-workspace-std' }