Make C strings safer, by encoding the absence of nulls in the type system. This is made possible by https://github.com/rust-lang/rust/pull/141001 which should hit stable in 1.89.

Performance seems largely unaffected. See details [for arm64](https://github.com/rust-lang/literal-escaper/pull/15#issuecomment-3004486202) and [for amd64](https://github.com/rust-lang/literal-escaper/pull/15#issuecomment-3004532601).