@hkBst 2025-06-23
New bench after rebase, still regressing:
| bench name | current | new | diff |
|-|-|-|-|
| bench_check_raw_byte_str_ascii | 44639.04 ns/iter (+/- 71.86) | 44633.22 ns/iter (+/- 65.53) | -0.0% |
| bench_check_raw_c_str_ascii | 41289.06 ns/iter (+/- 80.46) | 41280.67 ns/iter (+/- 52.0) | -0.0% |
| bench_check_raw_c_str_non_ascii | 52493.15 ns/iter (+/- 75.91) | 52491.51 ns/iter (+/- 81.16) | -0.0% |
| bench_check_raw_c_str_unicode | 174601.08 ns/iter (+/- 1708.93) | 175265.45 ns/iter (+/- 932.32) | 0.4% |
| bench_check_raw_str_ascii | 40617.39 ns/iter (+/- 72.24) | 40638.75 ns/iter (+/- 67.75) | 0.1% |
| bench_check_raw_str_non_ascii | 50443.77 ns/iter (+/- 64.5) | 50447.15 ns/iter (+/- 85.57) | 0.0% |
| bench_check_raw_str_unicode | 167877.1 ns/iter (+/- 450.94) | 167587.6 ns/iter (+/- 363.93) | -0.2% |
| bench_skip_ascii_whitespace | 13124.12 ns/iter (+/- 21.57) | 12092.16 ns/iter (+/- 7.34) | -7.9% |
| bench_unescape_byte_str_ascii | 72015.72 ns/iter (+/- 197.75) | 71917.45 ns/iter (+/- 159.28) | -0.1% |
| bench_unescape_byte_str_ascii_escape | 81212.75 ns/iter (+/- 125.82) | 101553.12 ns/iter (+/- 213.74) | 25.0% |
| bench_unescape_byte_str_hex_escape | 125661.74 ns/iter (+/- 205.48) | 145728.98 ns/iter (+/- 190.31) | 16.0% |
| bench_unescape_byte_str_mixed_escape | 347796.68 ns/iter (+/- 1137.38) | 375593.79 ns/iter (+/- 555.62) | 8.0% |
| bench_unescape_c_str_ascii | 81542.57 ns/iter (+/- 178.04) | 81887.19 ns/iter (+/- 357.76) | 0.4% |
| bench_unescape_c_str_ascii_escape | 96352.47 ns/iter (+/- 216.99) | 125572.16 ns/iter (+/- 210.0) | 30.3% |
| bench_unescape_c_str_hex_escape_ascii | 161409.62 ns/iter (+/- 347.51) | 186294.32 ns/iter (+/- 387.11) | 15.4% |
| bench_unescape_c_str_hex_escape_byte | 162213.98 ns/iter (+/- 266.46) | 186187.18 ns/iter (+/- 642.11) | 14.8% |
| bench_unescape_c_str_mixed_escape | 1044968.1 ns/iter (+/- 5590.8) | 1146097.5 ns/iter (+/- 3564.66) | 9.7% |
| bench_unescape_c_str_non_ascii | 110542.91 ns/iter (+/- 193.2) | 110230.1 ns/iter (+/- 207.58) | -0.3% |
| bench_unescape_c_str_unicode | 384345.8 ns/iter (+/- 1529.57) | 388469.95 ns/iter (+/- 1881.49) | 1.1% |
| bench_unescape_c_str_unicode_escape | 308380.05 ns/iter (+/- 1680.4) | 316255.67 ns/iter (+/- 657.98) | 2.6% |
| bench_unescape_str_ascii | 77109.62 ns/iter (+/- 134.09) | 71986.74 ns/iter (+/- 153.02) | -6.6% |
| bench_unescape_str_ascii_escape | 105410.17 ns/iter (+/- 245.33) | 121219.55 ns/iter (+/- 541.69) | 15.0% |
| bench_unescape_str_hex_escape | 165387.27 ns/iter (+/- 692.67) | 179651.5 ns/iter (+/- 463.31) | 8.6% |
| bench_unescape_str_mixed_escape | 905730.0 ns/iter (+/- 2516.37) | 968862.0 ns/iter (+/- 3614.17) | 7.0% |
| bench_unescape_str_non_ascii | 97985.04 ns/iter (+/- 204.93) | 98175.32 ns/iter (+/- 191.07) | 0.2% |
| bench_unescape_str_unicode | 343506.0 ns/iter (+/- 767.26) | 344200.25 ns/iter (+/- 713.67) | 0.2% |
| bench_unescape_str_unicode_escape | 596488.05 ns/iter (+/- 2449.86) | 618829.1 ns/iter (+/- 4882.74) | 3.7% |