@hkBst 2025-06-25
Today's nightly (28f1c8079 2025-06-24) bench rerun (aarch64):
| bench name | current | new | diff |
|-|-|-|-|
| bench_check_raw_byte_str_ascii | 44342.34 ns/iter (+/- 83.84) | 44553.86 ns/iter (+/- 64.83) | 0.5% |
| bench_check_raw_c_str_ascii | 41067.41 ns/iter (+/- 75.04) | 41255.44 ns/iter (+/- 92.79) | 0.5% |
| bench_check_raw_c_str_non_ascii | 52563.06 ns/iter (+/- 81.78) | 52455.53 ns/iter (+/- 103.22) | -0.2% |
| bench_check_raw_c_str_unicode | 173704.5 ns/iter (+/- 1072.52) | 175189.6 ns/iter (+/- 1263.83) | 0.9% |
| bench_check_raw_str_ascii | 39604.61 ns/iter (+/- 54.11) | 40623.23 ns/iter (+/- 73.56) | 2.6% |
| bench_check_raw_str_non_ascii | 52355.27 ns/iter (+/- 89.8) | 50412.34 ns/iter (+/- 173.49) | -3.7% |
| bench_check_raw_str_unicode | 167169.1 ns/iter (+/- 305.45) | 168095.96 ns/iter (+/- 548.09) | 0.6% |
| bench_skip_ascii_whitespace | 12080.81 ns/iter (+/- 36.06) | 12059.09 ns/iter (+/- 8.13) | -0.2% |
| bench_unescape_byte_str_ascii | 72210.47 ns/iter (+/- 104.8) | 69120.45 ns/iter (+/- 139.45) | -4.3% |
| bench_unescape_byte_str_ascii_escape | 81180.9 ns/iter (+/- 128.21) | 81455.81 ns/iter (+/- 199.26) | 0.3% |
| bench_unescape_byte_str_hex_escape | 125616.36 ns/iter (+/- 240.7) | 125825.2 ns/iter (+/- 278.53) | 0.2% |
| bench_unescape_byte_str_mixed_escape | 348300.4 ns/iter (+/- 2102.29) | 348917.15 ns/iter (+/- 1267.77) | 0.2% |
| bench_unescape_c_str_ascii | 79577.9 ns/iter (+/- 165.32) | 80953.54 ns/iter (+/- 144.05) | 1.7% |
| bench_unescape_c_str_ascii_escape | 100745.16 ns/iter (+/- 174.85) | 100731.54 ns/iter (+/- 273.16) | -0.0% |
| bench_unescape_c_str_hex_escape_ascii | 156802.03 ns/iter (+/- 375.58) | 153611.05 ns/iter (+/- 288.01) | -2.0% |
| bench_unescape_c_str_hex_escape_byte | 152797.83 ns/iter (+/- 316.83) | 153587.95 ns/iter (+/- 258.41) | 0.5% |
| bench_unescape_c_str_mixed_escape | 1032048.4 ns/iter (+/- 2571.04) | 1030947.4 ns/iter (+/- 3510.86) | -0.1% |
| bench_unescape_c_str_non_ascii | 98631.62 ns/iter (+/- 148.69) | 97823.5 ns/iter (+/- 995.74) | -0.8% |
| bench_unescape_c_str_unicode | 354322.95 ns/iter (+/- 1250.7) | 358356.0 ns/iter (+/- 2320.65) | 1.1% |
| bench_unescape_c_str_unicode_escape | 313363.97 ns/iter (+/- 586.1) | 310180.1 ns/iter (+/- 811.19) | -1.0% |
| bench_unescape_str_ascii | 73216.99 ns/iter (+/- 146.36) | 71598.75 ns/iter (+/- 182.65) | -2.2% |
| bench_unescape_str_ascii_escape | 102818.88 ns/iter (+/- 210.38) | 102694.37 ns/iter (+/- 225.52) | -0.1% |
| bench_unescape_str_hex_escape | 164356.28 ns/iter (+/- 249.14) | 165551.38 ns/iter (+/- 247.75) | 0.7% |
| bench_unescape_str_mixed_escape | 903389.0 ns/iter (+/- 1952.41) | 904361.1 ns/iter (+/- 2725.98) | 0.1% |
| bench_unescape_str_non_ascii | 98219.37 ns/iter (+/- 215.35) | 98119.87 ns/iter (+/- 195.89) | -0.1% |
| bench_unescape_str_unicode | 342872.35 ns/iter (+/- 1324.02) | 341494.5 ns/iter (+/- 661.5) | -0.4% |
| bench_unescape_str_unicode_escape | 594178.6 ns/iter (+/- 2321.63) | 595380.3 ns/iter (+/- 5439.95) | 0.2% |

 - bench_check_raw_str_ascii and bench_check_raw_str_non_ascii have both inverted (win<->loss) their results (I suppose this is what they call bimodal in the rustc perf reports)
 - bench_unescape_byte_str_ascii got worse in main
 - followed by more results that either stayed similar, inverted or got slightly better