@hkBst 2025-06-25
In conclusion it seems that the perf impact of this change is largely neutral.