@hkBst 2025-06-25
Well, the main reason to do it is to give more guarantees for C strings, which cannot contain nulls. This change makes this guaranteed by the type system.