@hkBst 2025-06-25
Sure waiting is one options, and given 1.88 is around the corner, it should only be about 6 weeks.

Possible alternatives:
 - a "nightly" feature flag, though this would I think involve a lot of temporary code duplication. Seems unwieldy.
 - an MSRV set to 1.89 for the next release? Seems simple, but is it effective?