@hkBst 2025-07-01
Would it help if we change the CI temporarily to 1.89 (current beta)?