@GuillaumeGomez 2025-07-01
Ok let's do it. Please also open an issue to change back the CI to stable once 1.89 is released.