@hkBst 2025-07-01
Added commit to change CI to beta and add issue #19 to reset it in 6 weeks.