@hkBst 2025-07-01
@Urgau performance seems largely unaffected by this change, after we added `#[inline]` to all small functions to prevent the regressions. I discuss the details in comments [for arm64](https://github.com/rust-lang/literal-escaper/pull/15#issuecomment-3004486202) and [for amd64](https://github.com/rust-lang/literal-escaper/pull/15#issuecomment-3004532601).

Can you explain your concern regarding rust analyzer? They can follow 6 weeks later, but I think they may also not have the same release trains, so could do a release using the new version much sooner than rustc. Is there a benefit to rustc also waiting 6 more weeks?